refactor(provider-setup): extract setupLanguageModelOpenAI to shared helper
Moves the duplicated setupLanguageModelOpenAI function from
memory-history-regression.spec.ts and loop-component-regression.spec.ts
into tests/helpers/provider-setup/setup-language-model-openai.ts.

Both specs now import from the shared location, eliminating the risk of
the two copies diverging when the provider modal UI changes.
