+++ tests/tests-automations/regression/core-components/loop-component-regression.spec.ts
@@ -2,6 +2,7 @@ import type { Page } from "@playwright/test";
 import { expect, test } from "../../../fixtures/fixtures";
 import { adjustScreenView } from "../../../helpers/ui/adjust-screen-view";
 import { awaitBootstrapTest } from "../../../helpers/other/await-bootstrap-test";
+import { setupLanguageModelOpenAI } from "../../../helpers/provider-setup/setup-language-model-openai";
 
 // Run tests serially to avoid "flow must be unique" 400 errors from parallel autosaves
 test.describe.configure({ mode: "serial" });
@@ -22,65 +23,6 @@ async function addLoopComponent(page: Page) {
   await page.waitForSelector('[data-testid="title-Loop"]', {
     timeout: 15000,
   });
-}
-
-// Configure the Language Model component's provider via "Setup Provider".
-// The Research Translation Loop template ships with an unconfigured Language Model
-// node — without this step the flow fails with "A model selection is required".
-// Mirrors the pattern from memory-history-regression.spec.ts.
-async function setupLanguageModelOpenAI(page: Page): Promise<void> {
-  const modelDropdown = page.getByTestId("model_model");
-  const hasModelDropdown = await modelDropdown.isVisible({ timeout: 5000 }).catch(() => false);
-
-  if (!hasModelDropdown) {
-    await page.getByRole("button", { name: "Setup Provider" }).click();
-    await page.waitForSelector('[data-testid="provider-item-OpenAI"]', { timeout: 10000 });
-    await page.getByTestId("provider-item-OpenAI").click();
-
-    const apiKeyInput = page.getByPlaceholder("sk-...");
-    await apiKeyInput.waitFor({ state: "visible", timeout: 5000 }).catch(() => {});
-
-    const apiKey = process.env.OPENAI_API_KEY ?? "";
-    if ((await apiKeyInput.count()) > 0 && apiKey) {
-      await apiKeyInput.click();
-      await apiKeyInput.pressSequentially(apiKey, { delay: 0 });
-
-      const saveBtn = page.getByRole("button", { name: "Save", exact: true });
-      const replaceBtn = page.getByRole("button", { name: "Replace", exact: true });
-
-      if ((await saveBtn.count()) > 0) {
-        await saveBtn.click();
-      } else if ((await replaceBtn.count()) > 0) {
-        await replaceBtn.click();
-      }
-
-      await replaceBtn.waitFor({ state: "visible", timeout: 30000 });
-      await page.locator('[data-testid^="llm-toggle"]').first()
-        .waitFor({ state: "visible", timeout: 15000 })
-        .catch(() => {});
-    }
-
-    const toggles = page.locator('[data-testid^="llm-toggle"]');
-    const toggleCount = await toggles.count();
-    for (let i = 0; i < toggleCount; i++) {
-      if ((await toggles.nth(i).getAttribute("aria-checked")) !== "true") {
-        await toggles.nth(i).click();
-      }
-    }
-
-    await page.keyboard.press("Escape");
-    await modelDropdown.waitFor({ state: "visible", timeout: 30000 });
-  }
-
-  await modelDropdown.click();
-  await page
-    .locator('[data-testid$="-option"]', { hasText: "gpt-4o-mini" })
-    .first()
-    .waitFor({ state: "visible", timeout: 10000 });
-  await page
-    .locator('[data-testid$="-option"]', { hasText: "gpt-4o-mini" })
-    .first()
-    .click();
 }
 
 // =============================================================================
