+++ tests/tests-automations/regression/core-functionality/llm-agents/memory-history-regression.spec.ts
@@ -5,6 +5,7 @@ import { expect, test } from "../../../../fixtures/fixtures";
 import { adjustScreenView } from "../../../../helpers/ui/adjust-screen-view";
 import { updateOldComponents } from "../../../../helpers/flows/update-old-components";
 import { PlaygroundPage } from "../../../../pages";
+import { setupLanguageModelOpenAI } from "../../../../helpers/provider-setup/setup-language-model-openai";
 
 if (!process.env.CI) {
   dotenv.config({ path: path.resolve(__dirname, "../../../../.env") });
@@ -41,64 +42,6 @@ async function loadMemoryChatbot(page: Page): Promise<void> {
   await adjustScreenView(page);
   await updateOldComponents(page);
   await adjustScreenView(page);
-}
-
-async function setupLanguageModelOpenAI(page: Page): Promise<void> {
-  const modelDropdown = page.getByTestId("model_model");
-  const hasModelDropdown = await modelDropdown.isVisible({ timeout: 5000 }).catch(() => false);
-
-  if (!hasModelDropdown) {
-    // "Setup Provider" opens the provider modal (no data-testid on this button)
-    await page.getByRole("button", { name: "Setup Provider" }).click();
-    await page.waitForSelector('[data-testid="provider-item-OpenAI"]', { timeout: 10000 });
-    await page.getByTestId("provider-item-OpenAI").click();
-
-    const apiKeyInput = page.getByPlaceholder("sk-...");
-    // Wait for the form panel to animate in before checking visibility
-    await apiKeyInput.waitFor({ state: "visible", timeout: 5000 }).catch(() => {});
-
-    const apiKey = process.env.OPENAI_API_KEY ?? "";
-    if ((await apiKeyInput.count()) > 0 && apiKey) {
-      await apiKeyInput.click();
-      await apiKeyInput.pressSequentially(apiKey, { delay: 0 });
-
-      const saveBtn = page.getByRole("button", { name: "Save", exact: true });
-      const replaceBtn = page.getByRole("button", { name: "Replace", exact: true });
-
-      if ((await saveBtn.count()) > 0) {
-        await saveBtn.click();
-      } else if ((await replaceBtn.count()) > 0) {
-        await replaceBtn.click();
-      }
-
-      // After save the button becomes "Replace" — wait for that to confirm save completed
-      await replaceBtn.waitFor({ state: "visible", timeout: 30000 });
-      // Wait for model toggles to load
-      await page.locator('[data-testid^="llm-toggle"]').first()
-        .waitFor({ state: "visible", timeout: 15000 })
-        .catch(() => {});
-    }
-
-    // Enable any model toggles that are off
-    const toggles = page.locator('[data-testid^="llm-toggle"]');
-    const toggleCount = await toggles.count();
-    for (let i = 0; i < toggleCount; i++) {
-      if ((await toggles.nth(i).getAttribute("aria-checked")) !== "true") {
-        await toggles.nth(i).click();
-      }
-    }
-
-    // Escape closes the Dialog and triggers refreshAllModelInputs on the node
-    await page.keyboard.press("Escape");
-    await modelDropdown.waitFor({ state: "visible", timeout: 30000 });
-  }
-
-  await modelDropdown.click();
-  await page
-    .locator('[data-testid$="-option"]', { hasText: "gpt-4o-mini" })
-    .first()
-    .waitFor({ state: "visible", timeout: 10000 });
-  await page.locator('[data-testid$="-option"]', { hasText: "gpt-4o-mini" }).first().click();
 }
 
 async function waitForChatResponse(page: Page): Promise<void> {
